Implement qasm include fill in
Does only work for includes in the root file, i.e. does not work
recursively at the moment.

The problem is that the included files are not valid QASM files
themselves. Hence they can't be parsed on their own (to find includes),
but they have to be first inserted into the main code, which can then be
parsed from start to end again. All includes would then be relative to
the main file though, and not to the included file.

A solution would be to store for the assembled code the source of each
block of text, such that when a new include is resolved, it can be done
relative to the file that the include was originally from.