Simplify AST considerably
Having only program, statement and expr makes like way easier.
Of course, this allows ASTs that would be invalid in terms of grammar.
However, we assume correct input for now, and even in the future
checking the semantic correctness (e.g. no measurements in gates) could
be done while translating to the AST, ensuring that the resulting AST is
valid.