Add structs for AST gate defs/calls
Enum types are not independent types. Hence, it is not possible to
create e.g. a list of Statement::GateDeclaration. Also, it's not
possible writing a function taking only this kind of statement as
argument. So a lot of casting would be required. While there is ongoing
development to make enum types as own types, the current way is to
define a struct for each enum value and just put the struct inside.
This change might also be done for the other statements at a later
point.