Add vec of tetra tensors as result of qasm import
The current solution is a bit inefficient in that it creates a new
tensor every time (it could create one and clone it instead).
However, it was planned to find a smarter way anyway (that doesn't
require having the same CX tensor ten times), so it doesn't make sense
to put too much effort into an efficient solution in this commit.