Use newer U gate definition
The old and new definitions only differ in global phase.
The new definition is clearer and easier to compute.
Furthermore, it matches with other definitons better:
E.g. the H gate is defined as U(pi/2, 0, pi). With the old definition,
this would be -i/sqrt(2) * [[1, 1], [1, -1]].
With the new definition, it is 1/sqrt(2) * [[1, 1], [1, -1]]. The later
is the usual textbook definition.