Change to FxHashMap for reproducibility
FxHashMap uses a deterministic hasher with a fixed seed, hence the hashes and thus the iteration order of HashMaps will be the same between program runs. This greatly helps in debugging. Also, since the Hasher is not designed to secure against DoS attacks, it can even be a bit faster.