Refactor partitioning and path finding into separate method
This allows the users of simulated annealing and the genetic algorithm to get the exact same path that achieved the best score.

Of course, the methods could also directly return this information. However, this would require keeping this information around for every candidate in a thread-safe manner which could be complicated. Just storing/passing Vec<usize> saves on headaches and reconstructing the solution is considerably fast.