Fix wrong partition recalculated in simulated annealing
Previously, the old partition was determined by looking up the partition of the randomly chosen tensor. Or at least that was the intention. `tensor_index` is wrongly named, because it actually corresponds to an index into the contraction path and not a tensor index. Hence, using this to access the partitioning didn't necessarily return the correct partition.
The solution: `partition_index` corresponds to the original partition, as can be seen e.g. in the filter of the `shifted_tensor` loop.