Remove special leaf tensor handling from push_tensor
While it is a nice idea to keep a flat hierarchy, this actually made code more complex than simple. For example, it can happen that KaHyPar only puts a single tensor in a partition. Then suddenly, one of the partition tensors is a leaf tensor and not a composite tensor. This requires extra handling in contraction path finding, MPI communication, ... .