Remove random_tensor method
The method wasn't used and wasn't useful: It took the number of dimensions and generated a tensor with random bond dims and legs from 0..n. So if one used this method to generate two tensors, they would share the same legs (but with different bond dimensions, which is actually invalid). So unless one wants to test serialization (but without data) or contraction to scalars, there's no good use for this method.