Decrease RandomGreedy candidates to 100
RandomGreedy is the main bottleneck of our simulated annealing algorithm
(like 98% of the run time). 500 trials are quite a lot. I tried with
100 trials and the results were similar; the mean was even better, but
the worst cases were a bit worse, which makes sense.
So using a lot of trials is good to counter worst cases, but probably
not worth on average. The run time is 2.5 times faster now.