Fix output permutation of statevector
The order of the dimensions of a tensor after contraction can be
arbitrary. This is fine if the result is a scalar, but if we want to
e.g. compute a statevector, we need some fixed order to know which
amplitude belongs to what bitstring. In the best case, we want the same
order as one would normally write it out ("000", "001", "010", "011",
...). Since the final order of legs is dependent on the contraction
order, the Circuit doesn't know the order and can't return a
Permutation. Hence this workaround with the Permutor, which can apply
the permutation once the tensor is actually contracted.